requirements.txt
python-telegram-bot
supabase
google-generativeai
python-dotenv
fastapi
uvicorn
